ResuMate -  Resume Parsing Assistant/requirements.txt
streamlit==1.31.0
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.0.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl
transformers==4.36.2
torch==2.2.0 